Added time.perf_counter() to get more precise time elapsed
Waiting for temperatures from the UDP socket fucks takes a while sometimes
because of weak wifi connection, taking more time than we would expect.
This translates into a quantity of added up time far bigger than
I would've expected and that is bad.
So, introducing time.perf_counter difference between start and end of loop
to get a preciser time elapsed to write in logs.

The only way not to make the UDP wait impact the loop frequency
would be to implement multiprocessing, and request temperatures
in a separate thread. But I really don't feel like doing that now